+++ jest.config.js
@@ -1,6 +1,7 @@
 module.exports = {
   collectCoverageFrom: ['src/**/*.js'],
   coverageThreshold: { global: { lines: 95 } },
+  restoreMocks: true,
   setupFilesAfterEnv: ['<rootDir>/jest.setup.js'],
   testRegex: '(/(test|__tests__)/(?!_).*|(\\.|/)(test|spec))\\.js$',
   testURL: 'http://localhost',