+++ src/theme/symbol.js
@@ -1,0 +1,2 @@
+// This symbol is for internal to skip name validation when creating theme.
+export default Symbol('skipThemeValidation')